Finished this feature
4. Ascension Drop Change:
Drop rate will now increase based on how many of that specific monster have been farmed since the last time that that type of monster dropped an Ascension Orb. For example, a slime will have a higher chance to drop an Ascension Orb if many slimes have been slain since the last time a slime dropped an orb. Each type of monster is tracked separately and that monster's counter is reset when it drops an orb for anyone.
An optional kill counter has been added to the experience ui. Toggle with /expkillcount (defaults to OFF). Shows the counter for the most recently killed monster type only